Fibonacci con int

+++ Intentos/app/src/main/java/com/example/intentos/Fibonacci.java
@@ -2,11 +2,11 @@ package com.example.intentos;
 
 public class Fibonacci {
 
-    public static String calcularFibonacci(long n) {
+    public static String calcularFibonacci(int n) {
         double sqrt5 = Math.sqrt(5);
         double phi = (1 + sqrt5) / 2;
         String result;
 
-        return result = String.valueOf((long) ((Math.pow(phi, n) - Math.pow(-phi, -n)) / sqrt5));
+        return result = String.valueOf((int) ((Math.pow(phi, n) - Math.pow(-phi, -n)) / sqrt5));
     }
 }
